Revert "arm64: Treat ESR_Condor as a 64-bit register"
This reverts commit 46ddfb9d1e462c88163a239a7882ece6d2b45ff1 which is
commit 8d56e5c5a99ce1d17d39ce5a8260e42c2a2d7682 upstream.

It broke the abi and the change that was needing it was reverted so this
is no longer needed as well.

Bug: 161946584

+++ arch/arm64/include/asm/system_misc.h
@@ -23,9 +23,9 @@ void die(const char *msg, struct pt_regs *regs, int err);
 struct siginfo;
 void arm64_notify_die(const char *str, struct pt_regs *regs,
 		      int signo, int sicode, unsigned long far,
-		      unsigned long err);
+		      int err);
 
-void hook_debug_fault_code(int nr, int (*fn)(unsigned long, unsigned long,
+void hook_debug_fault_code(int nr, int (*fn)(unsigned long, unsigned int,
 					     struct pt_regs *),
 			   int sig, int code, const char *name);
